[cli][functions] Replace @squoosh/lib with sharp.
