+++ packages/cli/src/types/global.d.ts
@@ -1,5 +1,2 @@
 // See: https://github.com/KhronosGroup/glTF-Validator/issues/114
 declare module 'gltf-validator';
-
-// See: https://github.com/GoogleChromeLabs/squoosh/issues/1211
-declare module '@squoosh/lib';
